optimize ram / flash usage

+++ src/activities/reader/EpubReaderActivity.cpp
@@ -749,10 +749,10 @@ void EpubReaderActivity::renderStatusBar(const int orientedMarginRight, const in
 
     // Hide percentage when progress bar is shown to reduce clutter
     if (showProgressPercentage) {
-      snprintf(progressStr, sizeof(progressStr), "%d/%d  %.0f%%", section->currentPage + 1, section->pageCount,
-               bookProgress);
-    } else if (showBookPercentage) {
-      snprintf(progressStr, sizeof(progressStr), "%.0f%%", bookProgress);
+      snprintf(progressStr, sizeof(progressStr), "%d/%d  %d%%", section->currentPage + 1, section->pageCount,
+               (int)bookProgress);
+      } else if (showBookPercentage) {
+      snprintf(progressStr, sizeof(progressStr), "%d%%", (int)bookProgress);
     } else {
       snprintf(progressStr, sizeof(progressStr), "%d/%d", section->currentPage + 1, section->pageCount);
     }
